hackathon-0-portfolio/pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
dependencies = [
  "google-auth", "google-auth-oauthlib", "google-api-python-client",
  "watchdog", "schedule", "groq", "playwright", "python-dotenv",
  "requests"
]
